Persona clases

+++ apback/src/main/java/com/apback/repository/DatosUsuarioRepository.java
@@ -6,6 +6,6 @@ import org.springframework.stereotype.Repository;
 import com.apback.model.DatosUsuario;
 
 @Repository
-public interface  DatosUsuarioRepository  extends JpaRepository<DatosUsuario, Long>{
+public interface  DatosUsuarioRepository  extends JpaRepository<DatosUsuario, Integer>{
 
 }
